refactor : renamed `app/states` to `app/stores`
